refactor: creating another class for handling entities of the current user
